# Enable faster cuBLAS path for torch.linalg.lstsq for batch of small matrices